requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
chat_exporter==2.6.1
config==0.5.1
dislash.py==1.4.9
disnake==2.9.0
Django==4.2.3
mojang==0.2.0
Pillow==9.5.0
Pillow==10.0.0
psycopg2_binary==2.9.6
Py_SPW==1.4.4
Requests==2.31.0
validators==0.20.0
yandex_music==2.1.1
yt_dlp==2023.3.4
ytmusicapi==1.1.0
